Changed types Long to Integer to correspond the types in the POJOs

+++ src/main/java/io/wzw/backend/data/dao/RoleDAO.java
@@ -9,7 +9,7 @@ import io.wzw.backend.data.model.Role;
  */
 public interface RoleDAO {
 	
-	public Role selectById(Long id);
+	public Role selectById(Integer id);
 
 	public List<Role> selectAll();
